HDF5AnnData

Implementation of an HDF5-backed AnnData object. This class provides an interface to a H5AD file and minimal data is stored in memory until it is requested by the user. It is primarily designed as an intermediate object when reading/writing H5AD files but can be useful for accessing parts of large files.

See AnnData-usage for details on creating and using AnnData objects.

Value

An HDF5AnnData object

See also

Super class

AbstractAnnData -> HDF5AnnData

Active bindings

X

See AnnData-usage

layers

See AnnData-usage

obsm

See AnnData-usage

varm

See AnnData-usage

obsp

See AnnData-usage

varp

See AnnData-usage

obs

See AnnData-usage

var

See AnnData-usage

obs_names

See AnnData-usage

var_names

See AnnData-usage

uns

See AnnData-usage

Methods

Public methods

HDF5AnnData$new()

HDF5AnnData$obs_keys()

HDF5AnnData$var_keys()

HDF5AnnData$layers_keys()

HDF5AnnData$obsm_keys()

HDF5AnnData$varm_keys()

HDF5AnnData$obsp_keys()

HDF5AnnData$varp_keys()

HDF5AnnData$uns_keys()

HDF5AnnData$n_obs()

HDF5AnnData$n_vars()

HDF5AnnData$open()

HDF5AnnData$close()

HDF5AnnData$as_InMemoryAnnData()

Inherited methods

AbstractAnnData$as_HDF5AnnData()

AbstractAnnData$as_ReticulateAnnData()

AbstractAnnData$as_Seurat()

AbstractAnnData$as_SingleCellExperiment()

AbstractAnnData$as_ZarrAnnData()

AbstractAnnData$print()

AbstractAnnData$shape()

AbstractAnnData$write_h5ad()

AbstractAnnData$write_zarr()

HDF5AnnData$new()

HDF5AnnData constructor

Usage

HDF5AnnData$new(
  file,
  X = NULL,
  obs = NULL,
  var = NULL,
  layers = NULL,
  obsm = NULL,
  varm = NULL,
  obsp = NULL,
  varp = NULL,
  uns = NULL,
  shape = NULL,
  mode = c("a", "r", "r+", "w", "w-", "x"),
  backed = FALSE,
  compression = c("none", "gzip", "lzf"),
  chunk_size = "auto"
)

Arguments

file

The file name (character) of the .h5ad file. If this file already exits, other arguments must be NULL.

X

See the X slot in AnnData-usage

obs

See the obs slot in AnnData-usage

var

See the var slot in AnnData-usage

layers

See the layers slot in AnnData-usage

obsm

See the obsm slot in AnnData-usage

varm

See the varm slot in AnnData-usage

obsp

See the obsp slot in AnnData-usage

varp

See the varp slot in AnnData-usage

uns

See the uns slot in AnnData-usage

shape

Shape tuple (e.g. c(n_obs, n_vars)). Can be provided if both X or obs and var are not provided.

mode

The mode to open the HDF5 file. See as_HDF5AnnData() for details

backed

Whether the object is disk backed. See as_HDF5AnnData() for details

compression

The compression algorithm to use. See as_HDF5AnnData() for details

chunk_size

The target chunk size in bytes. See as_HDF5AnnData() for details

Details

The constructor creates a new HDF5 AnnData interface object. This can either be used to either connect to an existing .h5ad file or to create a new one. If any additional slot arguments are set an existing file will be overwritten.

HDF5AnnData$obs_keys()

See AnnData-usage

Usage

HDF5AnnData$obs_keys()

HDF5AnnData$var_keys()

See AnnData-usage

Usage

HDF5AnnData$var_keys()

HDF5AnnData$layers_keys()

See AnnData-usage

Usage

HDF5AnnData$layers_keys()

HDF5AnnData$obsm_keys()

See AnnData-usage

Usage

HDF5AnnData$obsm_keys()

HDF5AnnData$varm_keys()

See AnnData-usage

Usage

HDF5AnnData$varm_keys()

HDF5AnnData$obsp_keys()

See AnnData-usage

Usage

HDF5AnnData$obsp_keys()

HDF5AnnData$varp_keys()

See AnnData-usage

Usage

HDF5AnnData$varp_keys()

HDF5AnnData$uns_keys()

See AnnData-usage

Usage

HDF5AnnData$uns_keys()

HDF5AnnData$n_obs()

See the n_obs field in AnnData-usage

Usage

HDF5AnnData$n_obs()

HDF5AnnData$n_vars()

See the n_vars field in AnnData-usage

Usage

HDF5AnnData$n_vars()

HDF5AnnData$open()

Open the HDF5 file handle

Usage

HDF5AnnData$open(readonly = FALSE)

Arguments

readonly

Whether to open in read-only mode

HDF5AnnData$close()

Close the HDF5 file handle

Usage

HDF5AnnData$close()

HDF5AnnData$as_InMemoryAnnData()

Convert to an InMemoryAnnData. Any backed (DelayedArray) slots are materialized into ordinary in-memory matrices: an in-memory object should not stay tied to an on-disk file. Slots are read eagerly (via base rhdf5) rather than carried over as DelayedArrays, which is also faster than realizing them.

Usage

HDF5AnnData$as_InMemoryAnnData()