ReticulateAnnData

Implementation of an AnnData object that wraps a Python anndata AnnData object using reticulate. This allows direct interaction with Python AnnData objects while maintaining the R interface. It is useful when you already have a Python AnnData or to access functionality that has not yet been implemented in anndataR.

See AnnData-usage for details on creating and using AnnData objects.

Value

A ReticulateAnnData object

See also

Super class

AbstractAnnData -> ReticulateAnnData

Active bindings

X

See AnnData-usage

layers

See AnnData-usage

obs

See AnnData-usage

var

See AnnData-usage

obs_names

See AnnData-usage

var_names

See AnnData-usage

obsm

See AnnData-usage

varm

See AnnData-usage

obsp

See AnnData-usage

varp

See AnnData-usage

uns

See AnnData-usage

Methods

Public methods

ReticulateAnnData$new()

ReticulateAnnData$n_obs()

ReticulateAnnData$n_vars()

ReticulateAnnData$py_anndata()

Inherited methods

AbstractAnnData$as_HDF5AnnData()

AbstractAnnData$as_InMemoryAnnData()

AbstractAnnData$as_ReticulateAnnData()

AbstractAnnData$as_Seurat()

AbstractAnnData$as_SingleCellExperiment()

AbstractAnnData$as_ZarrAnnData()

AbstractAnnData$layers_keys()

AbstractAnnData$obs_keys()

AbstractAnnData$obsm_keys()

AbstractAnnData$obsp_keys()

AbstractAnnData$print()

AbstractAnnData$shape()

AbstractAnnData$uns_keys()

AbstractAnnData$var_keys()

AbstractAnnData$varm_keys()

AbstractAnnData$varp_keys()

AbstractAnnData$write_h5ad()

AbstractAnnData$write_zarr()

ReticulateAnnData$new()

ReticulateAnnData constructor

Usage

ReticulateAnnData$new(
  py_anndata = NULL,
  X = NULL,
  obs = NULL,
  var = NULL,
  layers = NULL,
  obsm = NULL,
  varm = NULL,
  obsp = NULL,
  varp = NULL,
  uns = NULL,
  shape = NULL
)

Arguments

py_anndata

A Python AnnData object created using reticulate, or NULL to create a new empty Python AnnData object

X

See the X slot in AnnData-usage (only used if py_anndata is NULL)

obs

See the obs slot in AnnData-usage (only used if py_anndata is NULL)

var

See the var slot in AnnData-usage (only used if py_anndata is NULL)

layers

See the layers slot in AnnData-usage (only used if py_anndata is NULL)

obsm

See the obsm slot in AnnData-usage (only used if py_anndata is NULL)

varm

See the varm slot in AnnData-usage (only used if py_anndata is NULL)

obsp

See the obsp slot in AnnData-usage (only used if py_anndata is NULL)

varp

See the varp slot in AnnData-usage (only used if py_anndata is NULL)

uns

See the uns slot in AnnData-usage (only used if py_anndata is NULL)

shape

Shape tuple (e.g. c(n_obs, n_vars)). Can be provided if both X or obs and var are not provided. (only used if py_anndata is NULL)

Details

The constructor creates a new ReticulateAnnData interface object that wraps a Python AnnData object. If py_anndata is provided, it must be a valid Python AnnData object. If NULL, a new Python AnnData object will be created using the other provided arguments.

ReticulateAnnData$n_obs()

See the n_obs field in AnnData-usage

Usage

ReticulateAnnData$n_obs()

ReticulateAnnData$n_vars()

See the n_vars field in AnnData-usage

Usage

ReticulateAnnData$n_vars()

ReticulateAnnData$py_anndata()

Get the underlying Python AnnData object

Usage

ReticulateAnnData$py_anndata()

Returns

The Python AnnData object wrapped by this ReticulateAnnData