Convert a SingleCellExperiment object to an AnnData object

from_SingleCellExperiment() converts a SingleCellExperiment to an AnnData object.

Usage

from_SingleCellExperiment(
  sce,
  output_class = c("InMemory", "HDF5AnnData"),
  x_mapping = NULL,
  layers_mapping = NULL,
  obs_mapping = NULL,
  var_mapping = NULL,
  obsm_mapping = NULL,
  varm_mapping = NULL,
  obsp_mapping = NULL,
  varp_mapping = NULL,
  uns_mapping = NULL,
  ...
)

Arguments

sce

An object inheriting from SingleCellExperiment.

output_class

Name of the AnnData class. Must be one of "HDF5AnnData" or "InMemoryAnnData".

x_mapping

Name of the assay in sce to use as the X matrix in the AnnData object.

layers_mapping

A named vector or list mapping layer names in AnnData to assay names in the SingleCellExperiment object. Each name corresponds to the layer name in AnnData, and each value to the assay name in SCE.

obs_mapping

A named vector or list mapping column names in AnnData's obs to column names in SCE's colData. Each name corresponds to a column name in AnnData's obs, and each value to a column name in SCE's colData.

var_mapping

A named vector or list mapping column names in AnnData's var to column names in SCE's rowData. Each name corresponds to a column name in AnnData's var, and each value to a column name in SCE's rowData.

obsm_mapping

A named vector mapping obsm keys in AnnData to reducedDims names in SCE. Each name corresponds to a key in AnnData's obsm, and each value to the name of a reducedDim in SCE. Example: obsm_mapping = c(X_pca = "pca", X_umap = "umap").

varm_mapping

A named vector mapping varm keys in AnnData to reducedDims names in SCE. Each name corresponds to a key in AnnData's varm, and each value to the name of a reducedDim in SCE. Example: varm_mapping = c(PCs = "pca").

obsp_mapping

A named vector or list mapping obsp keys in AnnData to colPairs in the SingleCellExperiment object. Each name corresponds to a key in AnnData's obsp, and each value to a name in SCE's colPairs.

varp_mapping

A named vector or list mapping varp keys in AnnData to rowPairs in the SingleCellExperiment object. Each name corresponds to a key in AnnData's varp, and each value to a name in SCE's rowPairs.

uns_mapping

A named vector or list mapping uns keys in AnnData to metadata in the SingleCellExperiment object. Each name corresponds to a key in AnnData's uns, and each value to a name in SCE's metadata.

...

Additional arguments to pass to the generator function.

Value

from_SingleCellExperiment() returns an AnnData object (e.g., InMemoryAnnData) representing the content of sce.

Examples

## construct an AnnData object from a SingleCellExperiment
library(SingleCellExperiment)
#> Loading required package: SummarizedExperiment
#> Loading required package: MatrixGenerics
#> Loading required package: matrixStats
#> 
#> Attaching package: ‘MatrixGenerics’
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> Loading required package: GenomicRanges
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> Loading required package: generics
#> 
#> Attaching package: ‘generics’
#> The following objects are masked from ‘package:base’:
#> 
#>     as.difftime, as.factor, as.ordered, intersect, is.element, setdiff,
#>     setequal, union
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, is.unsorted, lapply,
#>     mapply, match, mget, order, paste, pmax, pmax.int, pmin, pmin.int,
#>     rank, rbind, rownames, sapply, saveRDS, table, tapply, unique,
#>     unsplit, which.max, which.min
#> Loading required package: S4Vectors
#> 
#> Attaching package: ‘S4Vectors’
#> The following object is masked from ‘package:utils’:
#> 
#>     findMatches
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> 
#> Attaching package: ‘IRanges’
#> The following object is masked from ‘package:sp’:
#> 
#>     %over%
#> Loading required package: GenomeInfoDb
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: ‘Biobase’
#> The following object is masked from ‘package:MatrixGenerics’:
#> 
#>     rowMedians
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     anyMissing, rowMedians
#> Warning: replacing previous import ‘S4Arrays::makeNindexFromArrayViewport’ by ‘DelayedArray::makeNindexFromArrayViewport’ when loading ‘SummarizedExperiment’
#> 
#> Attaching package: ‘SummarizedExperiment’
#> The following object is masked from ‘package:Seurat’:
#> 
#>     Assays
#> The following object is masked from ‘package:SeuratObject’:
#> 
#>     Assays
sce <- SingleCellExperiment(
  assays = list(counts = matrix(1:5, 5L, 3L)),
  colData = DataFrame(cell = 1:3, row.names = paste0("Cell", 1:3)),
  rowData = DataFrame(gene = 1:5, row.names = paste0("Gene", 1:5))
)
from_SingleCellExperiment(sce, "InMemory")
#> AnnData object with n_obs × n_vars = 3 × 5
#>     obs: 'cell'
#>     var: 'gene'
#>     layers: 'counts'