AbstractAnnData

Abstract R6 class representing an AnnData object. Defines the interface.

Active bindings

X

NULL or an observation x variable matrix (without dimnames) consistent with the number of rows of obs and var.

layers

The layers slot. Must be NULL or a named list with with all elements having the dimensions consistent with obs and var.

obs

A data.frame with columns containing information about observations. The number of rows of obs defines the observation dimension of the AnnData object.

var

A data.frame with columns containing information about variables. The number of rows of var defines the variable dimension of the AnnData object.

obs_names

Either NULL or a vector of unique identifiers used to identify each row of obs and to act as an index into the observation dimension of the AnnData object. For compatibility with R representations, obs_names should be a character vector.

var_names

Either NULL or a vector of unique identifiers used to identify each row of var and to act as an index into the variable dimension of the AnnData object.. For compatibility with R representations, var_names should be a character vector.

Methods

Public methods

AbstractAnnData$print()

AbstractAnnData$shape()

AbstractAnnData$n_obs()

AbstractAnnData$n_vars()

AbstractAnnData$obs_keys()

AbstractAnnData$var_keys()

AbstractAnnData$layers_keys()

AbstractAnnData$to_SingleCellExperiment()

AbstractAnnData$to_Seurat()

AbstractAnnData$to_InMemoryAnnData()

AbstractAnnData$clone()

Method print()

Print a summary of the AnnData object. print() methods should be implemented so that they are not computationally expensive.

Usage

AbstractAnnData$print(...)

Arguments

...

Optional arguments to print method.

Method shape()

Dimensions (observations x variables) of the AnnData object.

Usage

AbstractAnnData$shape()

Method n_obs()

Number of observations in the AnnData object.

Usage

AbstractAnnData$n_obs()

Method n_vars()

Number of variables in the AnnData object.

Usage

AbstractAnnData$n_vars()

Method obs_keys()

Keys ('column names') of obs.

Usage

AbstractAnnData$obs_keys()

Method var_keys()

Keys ('column names') of var.

Usage

AbstractAnnData$var_keys()

Method layers_keys()

Keys (element names) of layers.

Usage

AbstractAnnData$layers_keys()

Method to_SingleCellExperiment()

Convert to SingleCellExperiment

Usage

AbstractAnnData$to_SingleCellExperiment()

Method to_Seurat()

Convert to Seurat

Usage

AbstractAnnData$to_Seurat()

Method to_InMemoryAnnData()

Convert to an InMemory AnnData

Usage

AbstractAnnData$to_InMemoryAnnData()

Method clone()

The objects of this class are cloneable with this method.

Usage

AbstractAnnData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.