Abstract AnnData class

This class is an abstract representation of an AnnData object. It is intended to be used as a base class for concrete implementations of AnnData objects, such as InMemoryAnnData or HDF5AnnData.

See AnnData-usage for details on creating and using AnnData objects.

Value

An AbstractAnnData object

See also

Active bindings

X

See AnnData-usage

layers

See AnnData-usage

obs

See AnnData-usage

var

See AnnData-usage

obs_names

See AnnData-usage

var_names

See AnnData-usage

obsm

See AnnData-usage

varm

See AnnData-usage

obsp

See AnnData-usage

varp

See AnnData-usage

uns

See AnnData-usage

Methods

Public methods

AbstractAnnData$print()

AbstractAnnData$shape()

AbstractAnnData$n_obs()

AbstractAnnData$n_vars()

AbstractAnnData$obs_keys()

AbstractAnnData$var_keys()

AbstractAnnData$layers_keys()

AbstractAnnData$obsm_keys()

AbstractAnnData$varm_keys()

AbstractAnnData$obsp_keys()

AbstractAnnData$varp_keys()

AbstractAnnData$uns_keys()

AbstractAnnData$as_SingleCellExperiment()

AbstractAnnData$as_Seurat()

AbstractAnnData$as_InMemoryAnnData()

AbstractAnnData$as_ReticulateAnnData()

AbstractAnnData$as_HDF5AnnData()

AbstractAnnData$write_h5ad()

AbstractAnnData$clone()

Method print()

See AnnData-usage

Usage

AbstractAnnData$print(...)

Arguments

...

Optional arguments to print method

Method shape()

See AnnData-usage

Usage

AbstractAnnData$shape()

Method n_obs()

See AnnData-usage

Usage

AbstractAnnData$n_obs()

Method n_vars()

See AnnData-usage

Usage

AbstractAnnData$n_vars()

Method obs_keys()

See AnnData-usage

Usage

AbstractAnnData$obs_keys()

Method var_keys()

See AnnData-usage

Usage

AbstractAnnData$var_keys()

Method layers_keys()

See AnnData-usage

Usage

AbstractAnnData$layers_keys()

Method obsm_keys()

See AnnData-usage

Usage

AbstractAnnData$obsm_keys()

Method varm_keys()

See AnnData-usage

Usage

AbstractAnnData$varm_keys()

Method obsp_keys()

See AnnData-usage

Usage

AbstractAnnData$obsp_keys()

Method varp_keys()

See AnnData-usage

Usage

AbstractAnnData$varp_keys()

Method uns_keys()

See AnnData-usage

Usage

AbstractAnnData$uns_keys()

Method as_SingleCellExperiment()

Convert to SingleCellExperiment

See as_SingleCellExperiment() for more details on the conversion

Usage

AbstractAnnData$as_SingleCellExperiment(
  x_mapping = NULL,
  assays_mapping = TRUE,
  colData_mapping = TRUE,
  rowData_mapping = TRUE,
  reducedDims_mapping = TRUE,
  colPairs_mapping = TRUE,
  rowPairs_mapping = TRUE,
  metadata_mapping = TRUE
)

Arguments

x_mapping

See as_SingleCellExperiment()

assays_mapping

See as_SingleCellExperiment()

colData_mapping

See as_SingleCellExperiment()

rowData_mapping

See as_SingleCellExperiment()

reducedDims_mapping

See as_SingleCellExperiment()

colPairs_mapping

See as_SingleCellExperiment()

rowPairs_mapping

See as_SingleCellExperiment()

metadata_mapping

See as_SingleCellExperiment()

Returns

A SingleCellExperiment object

Method as_Seurat()

Convert to Seurat

See as_Seurat() for more details on the conversion

Usage

AbstractAnnData$as_Seurat(
  assay_name = "RNA",
  x_mapping = NULL,
  layers_mapping = TRUE,
  object_metadata_mapping = TRUE,
  assay_metadata_mapping = TRUE,
  reduction_mapping = TRUE,
  graph_mapping = TRUE,
  misc_mapping = TRUE
)

Arguments

assay_name

See as_Seurat()

x_mapping

See as_Seurat()

layers_mapping

See as_Seurat()

object_metadata_mapping

See as_Seurat()

assay_metadata_mapping

See as_Seurat()

reduction_mapping

See as_Seurat()

graph_mapping

See as_Seurat()

misc_mapping

See as_Seurat()

Returns

A Seurat object

Method as_InMemoryAnnData()

Convert to an InMemoryAnnData

See as_InMemoryAnnData() for more details on the conversion

Usage

AbstractAnnData$as_InMemoryAnnData()

Returns

An InMemoryAnnData object

Method as_ReticulateAnnData()

Convert to a ReticulateAnnData

See as_ReticulateAnnData() for more details on the conversion

Usage

AbstractAnnData$as_ReticulateAnnData()

Returns

A ReticulateAnnData object

Method as_HDF5AnnData()

Convert to an HDF5AnnData

See as_HDF5AnnData() for more details on the conversion

Usage

AbstractAnnData$as_HDF5AnnData(
  file,
  compression = c("none", "gzip", "lzf"),
  chunk_size = "auto",
  mode = c("w-", "r", "r+", "a", "w", "x")
)

Arguments

file

See as_HDF5AnnData()

compression

See as_HDF5AnnData()

chunk_size

See as_HDF5AnnData()

mode

See as_HDF5AnnData()

Returns

An HDF5AnnData object

Method write_h5ad()

Write the AnnData object to an H5AD file

See write_h5ad() for details

Usage

AbstractAnnData$write_h5ad(
  path,
  compression = c("none", "gzip", "lzf"),
  chunk_size = "auto",
  mode = c("w-", "r", "r+", "a", "w", "x")
)

Arguments

path

See write_h5ad()

compression

See write_h5ad()

chunk_size

See write_h5ad()

mode

See write_h5ad()

Returns

path invisibly

Method clone()

The objects of this class are cloneable with this method.

Usage

AbstractAnnData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.