An in-memory AnnData object

This class is used to represent an AnnData object in memory. AnnData stores a data matrix X together with annotations of observations obs (obsm, obsp), variables var (varm, varp), and unstructured annotations uns.

To read an AnnData file from disk, use read_h5ad() instead.

Usage

AnnData(
  obs_names = NULL,
  var_names = NULL,
  X = NULL,
  obs = NULL,
  var = NULL,
  layers = NULL,
  obsm = NULL,
  varm = NULL,
  obsp = NULL,
  varp = NULL
)

Arguments

obs_names

A vector of unique identifiers used to identify each row of obs and to act as an index into the observation dimension of the AnnData object. The length of obs_names defines the observation dimension of the AnnData object.

var_names

A vector of unique identifiers used to identify each row of var and to act as an index into the variable dimension of the AnnData object. The length of var_names defines the variable dimension of the AnnData object.

X

Either NULL or a observation × variable matrix with dimensions consistent with obs and var.

obs

Either NULL or a data.frame with columns containing information about observations. If NULL, an n_obs×0 data frame will automatically be generated.

var

Either NULL or a data.frame with columns containing information about variables. If NULL, an n_vars×0 data frame will automatically be generated.

layers

Either NULL or a named list, where each element is an observation × variable matrix with dimensions consistent with obs and var.

obsm

The obsm slot is used to store multi-dimensional annotation arrays. It must be either NULL or a named list, where each element is a matrix with n_obs rows and an arbitrary number of columns.

varm

The varm slot is used to store multi-dimensional annotation arrays. It must be either NULL or a named list, where each element is a matrix with n_vars rows and an arbitrary number of columns.

obsp

The obsp slot is used to store sparse multi-dimensional annotation arrays. It must be either NULL or a named list, where each element is a sparse matrix where each dimension has length n_obs.

varp

The varp slot is used to store sparse multi-dimensional annotation arrays. It must be either NULL or a named list, where each element is a sparse matrix where each dimension has length n_vars.

Examples

adata <- AnnData(
  obs_names = paste0("obs", 1:3),
  var_names = paste0("var", 1:4),
  X = matrix(1:12, nrow = 3, ncol = 4),
  obs = data.frame(
    n_counts = c(1, 2, 3),
    n_cells = c(1, 2, 3)
  ),
  var = data.frame(
    n_cells = c(1, 2, 3, 4)
  )
)

adata
#> AnnData object with n_obs × n_vars = 3 × 4
#>     obs: 'n_counts', 'n_cells'
#>     var: 'n_cells'