HDF5AnnData

Implementation of an in memory AnnData object.

Super class

anndataR::AbstractAnnData -> HDF5AnnData

Active bindings

X

The X slot

layers

The layers slot. Must be NULL or a named list with with all elements having the dimensions consistent with obs and var.

obs

The obs slot

var

The var slot

obs_names

Names of observations

var_names

Names of variables

Methods

Public methods

HDF5AnnData$new()

HDF5AnnData$n_obs()

HDF5AnnData$n_vars()

HDF5AnnData$clone()

Inherited methods

anndataR::AbstractAnnData$layers_keys()

anndataR::AbstractAnnData$obs_keys()

anndataR::AbstractAnnData$print()

anndataR::AbstractAnnData$shape()

anndataR::AbstractAnnData$to_InMemoryAnnData()

anndataR::AbstractAnnData$to_Seurat()

anndataR::AbstractAnnData$to_SingleCellExperiment()

anndataR::AbstractAnnData$var_keys()

Method new()

HDF5AnnData constructor

Usage

HDF5AnnData$new(file, X, obs, var, obs_names, var_names, layers)

Arguments

file

The filename (character) of the .h5ad file. Alternatively, you can also provide an object created by [rhdf5::H5Fopen()]. If this file does not exist yet, you must provide values for X, obs, var to create a new AnnData with.

X

Either NULL or a observation x variable matrix with dimensions consistent with obs and var.

obs

A data.frame with columns containing information about observations. The number of rows of obs defines the observation dimension of the AnnData object.

var

A data.frame with columns containing information about variables. The number of rows of var defines the variable dimension of the AnnData object.

obs_names

Either NULL or a vector of unique identifiers used to identify each row of obs and to act as an index into the observation dimension of the AnnData object. For compatibility with R representations, obs_names should be a character vector.

var_names

Either NULL or a vector of unique identifers used to identify each row of var and to act as an index into the variable dimension of the AnnData object. For compatibility with R representations, var_names should be a character vector.

layers

Either NULL or a named list, where each element is an observation x variable matrix with dimensions consistent with obs and var.

Method n_obs()

Number of observations in the AnnData object

Usage

HDF5AnnData$n_obs()

Method n_vars()

Number of variables in the AnnData object

Usage

HDF5AnnData$n_vars()

Method clone()

The objects of this class are cloneable with this method.

Usage

HDF5AnnData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.