InMemoryAnnData

Implementation of an in memory AnnData object.

Super class

anndataR::AbstractAnnData -> InMemoryAnnData

Active bindings

X

NULL or an observation x variable matrix (without dimnames) consistent with the number of rows of obs and var.

layers

NULL or a named list with all elements having the dimensions consistent with obs and var.

obs

A data.frame with columns containing information about observations. The number of rows of obs defines the observation dimension of the AnnData object.

var

A data.frame with columns containing information about variables. The number of rows of var defines the variable dimension of the AnnData object.

obs_names

Either NULL or a vector of unique identifiers used to identify each row of obs and to act as an index into the observation dimension of the AnnData object. For compatibility with R representations, obs_names should be a character vector.

var_names

Either NULL or a vector of unique identifiers used to identify each row of var and to act as an index into the variable dimension of the AnnData object. For compatibility with R representations, var_names should be a character vector.

Methods

Public methods

InMemoryAnnData$new()

InMemoryAnnData$clone()

Inherited methods

anndataR::AbstractAnnData$layers_keys()

anndataR::AbstractAnnData$n_obs()

anndataR::AbstractAnnData$n_vars()

anndataR::AbstractAnnData$obs_keys()

anndataR::AbstractAnnData$print()

anndataR::AbstractAnnData$shape()

anndataR::AbstractAnnData$to_InMemoryAnnData()

anndataR::AbstractAnnData$to_Seurat()

anndataR::AbstractAnnData$to_SingleCellExperiment()

anndataR::AbstractAnnData$var_keys()

Method new()

Creates a new instance of an in memory AnnData object. Inherits from AbstractAnnData.

Usage

InMemoryAnnData$new(
  obs_names,
  var_names,
  X = NULL,
  obs = NULL,
  var = NULL,
  layers = NULL
)

Arguments

obs_names

A vector of unique identifiers used to identify each row of obs and to act as an index into the observation dimension of the AnnData object. The length of the obs_names defines the observation dimension of the AnnData object.

var_names

A vector of unique identifers used to identify each row of var and to act as an index into the variable dimension of the AnnData object. The length of the var_names defines the variable dimension of the AnnData object.

X

Either NULL or a observation × variable matrix with dimensions consistent with obs and var.

obs

Either NULL or a data.frame with columns containing information about observations. If NULL, an n_obs×0 data frame will automatically be generated.

var

Either NULL or a data.frame with columns containing information about variables. If NULL, an n_vars×0 data frame will automatically be generated.

layers

Either NULL or a named list, where each element is an observation × variable matrix with dimensions consistent with obs and var.

Method clone()

The objects of this class are cloneable with this method.

Usage

InMemoryAnnData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## complete example
ad <- InMemoryAnnData$new(
  X = matrix(1:15, 3L, 5L),
  layers = list(
    A = matrix(5:1, 3L, 5L),
    B = matrix(letters[1:5], 3L, 5L)
  ),
  obs = data.frame(cell = 1:3),
  var = data.frame(gene = 1:5),
  obs_names = LETTERS[1:3],
  var_names = letters[1:5]
)
ad
#> class: InMemoryAnnData
#> dim: 3 obs x 5 var
#> X: matrix
#> layers: A B
#> obs: cell
#> var: gene

## minimum example
ad <- InMemoryAnnData$new(
  obs_names = letters[1:10],
  var_names = LETTERS[1:5]
)
ad
#> class: InMemoryAnnData
#> dim: 10 obs x 5 var
#> X: 
#> layers:
#> obs:
#> var: