anndataR: AnnData interoperability in R

Bring the power and flexibility of AnnData to the R ecosystem, allowing you to effortlessly manipulate and analyse your single-cell data. This package lets you work with backed h5ad and zarr files, directly access various slots (e.g. X, obs, var), or convert the data into SingleCellExperiment and Seurat objects.

See also

Author

Maintainer: Robrecht Cannoodt rcannood@gmail.com (ORCID) (github: rcannood)

Authors:

Robrecht Cannoodt rcannood@gmail.com (ORCID) (github: rcannood)

Luke Zappia luke@lazappi.id.au (ORCID) (github: lazappi)

Martin Morgan mtmorgan.bioc@gmail.com (ORCID) (github: mtmorgan)

Louise Deconinck louise.deconinck@gmail.com (ORCID) (github: LouiseDck)

Other contributors:

Danila Bredikhin danila.bredikhin@embl.de (ORCID) (github: gtca) [contributor]

Isaac Virshup (ORCID) (github: ivirshup) [contributor]

Brian Schilder brian_schilder@alumni.brown.edu (ORCID) (github: bschilder) [contributor]

Chananchida Sang-aram (ORCID) (github: csangara) [contributor]

Artür Manukyan (ORCID) (github: Artur-man) [contributor]

Hugo Gruson (ORCID) (github: Bisaloo) [contributor]

Data Intuitive info@data-intuitive.com [funder]

Chan Zuckerberg Initiative [funder]

scverse consortium [sponsor]